lab3/przykladowe/Przykład 2/requirements.txt
msl-loadlib
matplotlib